requirements.txt
agno
uvicorn
python-dotenv